requirements.txt
attrs==15.2.0
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
BestBuyAPI==0.0.51
blinker==1.3
chardet==2.3.0
configglue==1.1.2
configobj==5.0.6
cryptography==1.2.3
dj-database-url==0.4.2
Django==1.9
django-cors-headers==2.0.2
django-extra-fields==0.9
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
ebaysdk==2.1.4
ecdsa==0.13
enum34==1.1.2
feedparser==5.1.3
gunicorn==19.7.1
httplib2==0.9.1
idna==2.0
ipaddress==1.0.16
lockfile==0.12.2
lxml==3.7.3
Mako==1.0.3
oauthlib==1.0.3
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.1
ptyprocess==0.5
reportlab==3.3.0
requests==2.7.0
service-identity==16.0.0
six==1.10.0
Twisted==16.0.0
urllib3==1.13.1
virtualenv==15.1.0
wapy==1.0.0
whitenoise==3.3.0
python-amazon-product-api==0.2.8
